refactor: make Ranking component
MyPage에 있던 Ranking을 컴포넌트로 만듬
