+++ front/src/pages/MyPage/MyPage.js
@@ -1,40 +1,22 @@
-import React, { useContext } from 'react';
-import { ThemeContext } from 'styled-components';
-import Tab from 'components/globalComponents/Tab/Tab';
-import TabPane from 'components/globalComponents/Tab/TabPane/TabPane';
-import RankingAll from 'components/Ranking/RankingAll/RankingAll';
-import RankingFriends from 'components/Ranking/RankingFriends/RankingFriends';
+import React from 'react';
 import MyInfoSection from 'components/MyInfoSection/MyInfoSection';
 import ContentSection from 'components/globalComponents/ContentSection/ContentSection';
 import Background from 'components/globalComponents/Container/Background.style';
+import Ranking from 'components/Ranking/Ranking';
 
-const MyPage = () => {
-  const themeContext = useContext(ThemeContext);
+const MyPage = () => (
+  <>
+    {/* <NavigationBar /> */}
+    <Background id="MyPage">
+      <ContentSection title="랭킹">
+        <Ranking />
+      </ContentSection>
 
-  return (
-    <>
-      {/* <NavigationBar /> */}
-      <Background id="MyPage">
-        <ContentSection title="랭킹">
-          <Tab
-            activeTabColor={themeContext.amour}
-            inActiveTabColor={themeContext.wePeep}
-          >
-            <TabPane paneName="전체 랭킹">
-              <RankingAll />
-            </TabPane>
-            <TabPane paneName="친구 랭킹">
-              <RankingFriends />
-            </TabPane>
-          </Tab>
-        </ContentSection>
-
-        <ContentSection title="내 정보">
-          <MyInfoSection />
-        </ContentSection>
-      </Background>
-    </>
-  );
-};
+      <ContentSection title="내 정보">
+        <MyInfoSection />
+      </ContentSection>
+    </Background>
+  </>
+);
 
 export default MyPage;
